+++ internal/services/system/utils.go
@@ -43,17 +43,3 @@ func NewFixedClock(t time.Time) *FixedClock {
 	return &FixedClock{fixed: t}
 }
 
-// Ptr returns a pointer to the given value.
-func Ptr[T any](v T) *T {
-	return &v
-}
-
-// StringPtr returns a pointer to the given string.
-func StringPtr(s string) *string {
-	return Ptr(s)
-}
-
-// IntPtr returns a pointer to the given int.
-func IntPtr(i int) *int {
-	return Ptr(i)
-}